Issues-53: Moved LikeButton example to separate module.
