+++ likebutton/src/main/java/com/example/likebutton/LikeButtonActivity.kt
@@ -1,0 +1,13 @@
+package com.example.likebutton
+
+import android.os.Bundle
+import android.support.v7.app.AppCompatActivity
+import com.kirchhoff.likebutton.R
+
+class LikeButtonActivity : AppCompatActivity() {
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.a_like_button)
+    }
+}
