+++ likebutton/src/main/java/com/example/likebutton/utils/MathUtils.java
@@ -1,0 +1,12 @@
+package com.example.likebutton.utils;
+
+public class MathUtils {
+
+    public static double mapValueFromRangeToRange(double value, double fromLow, double fromHigh, double toLow, double toHigh) {
+        return toLow + ((value - fromLow) / (fromHigh - fromLow) * (toHigh - toLow));
+    }
+
+    public static double clamp(double value, double low, double high) {
+        return Math.min(Math.max(value, low), high);
+    }
+}
