+++ likebutton/src/main/res/layout/a_like_button.xml
@@ -4,9 +4,10 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:gravity="center"
+    android:background="@color/black"
     android:orientation="vertical">
 
-    <com.example.kirchhoff.example.ui.view.view.LikeButtonView
+    <com.example.likebutton.view.LikeButtonView
         android:layout_width="200dp"
         android:layout_height="200dp"
         android:layout_gravity="center" />
