+++ likebutton/src/main/res/layout/v_like_button.xml
@@ -3,13 +3,13 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent">
 
-    <com.example.kirchhoff.example.ui.view.view.like.DotsView
+    <com.example.likebutton.view.DotsView
         android:id="@+id/vDotsView"
         android:layout_width="200dp"
         android:layout_height="200dp"
         android:layout_gravity="center" />
 
-    <com.example.kirchhoff.example.ui.view.view.like.CircleView
+    <com.example.likebutton.view.CircleView
         android:id="@+id/vCircle"
         android:layout_width="80dp"
         android:layout_height="80dp"
