+++ settings.gradle
@@ -1,1 +1,1 @@
-include ':app', ':curvedbottombar', ':viewpagerwithoutfragment', ':fabanimation', ':fastrecyclerview', ':recyclermvp', ':coordinatorbehavior', ':countdownlabel', ':recyclersearch', ':simpletransition'
+include ':app', ':curvedbottombar', ':viewpagerwithoutfragment', ':fabanimation', ':fastrecyclerview', ':recyclermvp', ':coordinatorbehavior', ':countdownlabel', ':recyclersearch', ':simpletransition', ':likebutton'
